feat(dependencies): add video processing support and improve error handling
This commit introduces a new optional dependency for video processing by adding av to the pyproject.toml file. It also updates the handling of the VideoClips class in video.py to provide a dummy implementation when the av module is not available, along with an informative error message guiding users on how to install the required package. These changes enhance the functionality and user experience for video-related tasks in Anomalib.

Signed-off-by: Samet Akcay <samet.akcay@intel.com>
Signed-off-by: samet-akcay <samet.akcay@intel.com>